Fix User-Agent per RFC7231
The User-Agent field has a particular expected format and the current
setting doesn't follow it. This commit fixes it and provides the
Nerves Hub Link (NHL) version - actually this library's version.
Assuming that there is any reason to debug HTTP downloads based on this
field, that seems like the most reasonable thing to check.